chore(components/landingpage): rename path
